+++ packer-vmis/main.tf
@@ -2,32 +2,26 @@
 
 resource "null_resource" "packer-build" {
   triggers = {
-    virtual_machine_image = var.virtual_machine_image
-    always_run            = "${timestamp()}"
+    distribution_name            = var.distribution_name
+    virtual_machine_image        = var.virtual_machine_image
+    virtual_machine_image_sha    = var.virtual_machine_image_sha
+    virtual_machine_image_exists = fileexists(".packer-vmis/${var.virtual_machine_image}.box")
+    # always_run            = "${timestamp()}"
   }
   provisioner "local-exec" {
     working_dir = "."
     command     = <<EOF
-RED='\033[0;31m' # Red Text
-GREEN='\033[0;32m' # Green Text
-BLUE='\033[0;34m' # Blue Text
-NC='\033[0m' # No Color
-
 cd packer-vmis
-packer build ubuntu-22.10-live-server-amd64.pkr.hcl
-
-if [ $? -eq 0 ]; then
-  printf "\n $GREEN Packer Succeeded $NC \n"
-else
-  printf "\n $RED Packer Failed $NC \n" >&2
-  exit 1
-fi
+packer build -var "distribution_name=${self.triggers.distribution_name}" \
+             -var "virtual_machine_image=${self.triggers.virtual_machine_image}" \
+             -var "virtual_machine_image_sha=${self.triggers.virtual_machine_image_sha}" \
+             ${self.triggers.virtual_machine_image}.pkr.hcl
 EOF
   }
 
-  # # Keep the image to avoid rebuilding each time; uncomment when stable.
-  # provisioner "local-exec" {
-  #   when    = destroy
-  #   command = "rm -rf ./packer-vmis/${var.virtual_machine_image}.box"
-  # }
+  # Keep the image to avoid rebuilding each time; uncomment when stable.
+  provisioner "local-exec" {
+    when    = destroy
+    command = "rm -rf ./packer-vmis/${self.triggers.virtual_machine_image}.box"
+  }
 }